S3 Interaction

AWSS3.s3_arn — Function

s3_arn(resource)
s3_arn(bucket,path)

Amazon Resource Name for S3 resource or bucket and path.

AWSS3.s3_get — Function

s3_get([::AbstractAWSConfig], bucket, path; <keyword arguments>)

Get Object from path in bucket.

Optional Arguments

version=: version of object to get.

retry=true: try again on "NoSuchBucket", "NoSuchKey" (common if object was recently created).

raw=false: return response as Vector{UInt8}

byte_range=nothing: given an iterator of (start_byte, end_byte) gets only the range of bytes of the object from start_byte to end_byte. For example, byte_range=1:4 gets bytes 1 to 4 inclusive. Arguments should use the Julia convention of 1-based indexing.

header::Dict{String,String}: pass in an HTTP header to the request.

As an example of how to set custom HTTP headers, the below is equivalent to s3_get(aws, bucket, path; byte_range=range):

s3_get(aws, bucket, path; headers=Dict{String,String}("Range" => "bytes=$(first(range)-1)-$(last(range)-1)"))

AWSS3.s3_get_file — Function

s3_get_file([::AbstractAWSConfig], bucket, path, filename; [version=], kwargs...)

Like s3_get but streams result directly to filename. Keyword arguments accept are the same as those for s3_get.

AWSS3.s3_get_meta — Function

s3getmeta([::AbstractAWSConfig], bucket, path; [version=], kwargs...)

HEAD Object

Retrieves metadata from an object without returning the object itself.

AWSS3.s3_exists — Function

s3_exists([::AbstractAWSConfig], bucket, path; version=nothing)

Returns a boolean whether an object exists at path in bucket.

Note that the AWS API's support for object versioning is quite limited and this check will involve try catch logic if version is not nothing.

AWSS3.s3_delete — Function

s3_delete([::AbstractAWSConfig], bucket, path; [version=], kwargs...)

DELETE Object

AWSS3.s3_copy — Function

s3_copy([::AbstractAWSConfig], bucket, path; to_bucket=bucket, to_path=path, kwargs...)

PUT Object - Copy

Optional Arguments

metadata::Dict=; optional x-amz-meta- headers.

AWSS3.s3_create_bucket — Function

s3_create_bucket([:AbstractAWSConfig], bucket; kwargs...)

PUT Bucket

AWSS3.s3_put_cors — Function

s3_put_cors([::AbstractAWSConfig], bucket, cors_config; kwargs...)

PUT Bucket cors

s3_put_cors("my_bucket", """
    <?xml version="1.0" encoding="UTF-8"?>
    <CORSConfiguration xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
        <CORSRule>
            <AllowedOrigin>http://my.domain.com</AllowedOrigin>
            <AllowedOrigin>http://my.other.domain.com</AllowedOrigin>
            <AllowedMethod>GET</AllowedMethod>
            <AllowedMethod>HEAD</AllowedMethod>
            <AllowedHeader>*</AllowedHeader>
            <ExposeHeader>Content-Range</ExposeHeader>
        </CORSRule>
    </CORSConfiguration>
"""

AWSS3.s3_enable_versioning — Function

s3_enable_versioning([::AbstractAWSConfig], bucket; kwargs...)

Enable versioning for bucket.

AWSS3.s3_put_tags — Function

s3_put_tags([::AbstractAWSConfig], bucket, [path,] tags::Dict; kwargs...)

PUT tags on bucket or object (path).

See also tags= option on s3_put.

AWSS3.s3_get_tags — Function

s3_get_tags([::AbstractAWSConfig], bucket, [path]; kwargs...)

Get tags from bucket or object (path).

AWSS3.s3_delete_tags — Function

s3_delete_tags([::AbstractAWSConfig], bucket, [path])

Delete tags from bucket or object (path).

AWSS3.s3_delete_bucket — Function

s3_delete_bucket([::AbstractAWSConfig], "bucket"; kwargs...)

DELETE Bucket.

AWSS3.s3_list_buckets — Function

s3_list_buckets([::AbstractAWSConfig]; kwargs...)

List of all buckets owned by the sender of the request.

AWSS3.s3_list_objects — Function

s3_list_objects([::AbstractAWSConfig], bucket, [path_prefix]; delimiter="/", max_items=1000, kwargs...)

List Objects in bucket with optional path_prefix.

Returns an iterator of Dicts with keys Key, LastModified, ETag, Size, Owner, StorageClass.

AWSS3.s3_list_keys — Function

s3_list_keys([::AbstractAWSConfig], bucket, [path_prefix]; kwargs...)

Like s3_list_objects but returns object keys as Vector{String}.

AWSS3.s3_purge_versions — Function

s3_purge_versions([::AbstractAWSConfig], bucket, [path [, pattern]]; kwargs...)

DELETE all object versions except for the latest version.

AWSS3.s3_put — Function

s3_put([::AbstractAWSConfig], bucket, path, data, data_type="", encoding=""; <keyword arguments>)

PUT Object data at path in bucket.

Optional Arguments

data_type=; Content-Type header.

encoding=; Content-Encoding header.

acl=; x-amz-acl header for setting access permissions with canned config. See here.

metadata::Dict=; x-amz-meta- headers.

tags::Dict=; x-amz-tagging- headers (see also s3_put_tags and s3_get_tags).

AWSS3.s3_sign_url — Function

s3_sign_url([::AbstractAWSConfig], bucket, path, [seconds=3600];
            [verb="GET"], [content_type="application/octet-stream"],
            [protocol="http"], [signature_version="v4"])

Create a pre-signed url for bucket and path (expires after for seconds).

To create an upload URL use verb="PUT" and set content_type to match the type used in the Content-Type header of the PUT request.

For compatibility, the signature version 2 signing process can be used by setting signature_version="v2" but it is recommended that the default version 4 is used.

url = s3_sign_url("my_bucket", "my_file.txt"; verb="PUT")
Requests.put(URI(url), "Hello!")

url = s3_sign_url("my_bucket", "my_file.txt";
                  verb="PUT", content_type="text/plain")

Requests.put(URI(url), "Hello!";
             headers=Dict("Content-Type" => "text/plain"))

AWSS3.s3_nuke_bucket — Function

s3_nuke_bucket(bucket_name)

This function is NOT exported on purpose. AWS does not officially support this type of action, although it is a very nice utility one this is not exported just as a safe measure against accidentally blowing up your bucket.

Warning

It will delete all versions of objects in the given bucket and then the bucket itself.

S3Path

Note that S3Path implements the AbstractPath interface, some the FilePathsBase documentation for the interface here.

AWSS3.S3Path — Type

S3Path()
S3Path(str; version::Union{Nothing, AbstractString}=nothing, config::Union{Nothing, AWS.AbstractAWSConfig}=nothing)

Construct a new AWS S3 path type which should be of the form "s3://<bucket>/prefix/to/my/object".

NOTES:

Directories are required to have a trailing / due to how S3 distinguishes files from folders, as internally they're just keys to objects.

Objects p"s3://bucket/a" and p"s3://bucket/a/b" can co-exist. If both of these objects exist listing the keys for p"s3://bucket/a" returns [p"s3://bucket/a"] while p"s3://bucket/a/" returns [p"s3://bucket/a/b"].

The drive property will return "s3://<bucket>"

On top of the standard path properties (e.g., segments, root, drive, separator), S3Paths also support bucket and key properties for your convenience.

If version argument is nothing, will return latest version of object. Version can be provided via either kwarg version or as suffix "?versionId=<object_version>" of str, e.g., "s3://<bucket>/prefix/to/my/object?versionId=<object_version>".

If config is left at its default value of nothing, then the latest global_aws_config() will be used in any operations involving the path. To "freeze" the config at construction time, explicitly pass an AbstractAWSConfig to the config keyword argument.

Base.stat — Function

stat(fp::S3Path)

Return the status struct for the S3 path analogously to stat for local directories.

Note that this cannot be used on a directory. This is because S3 is a pure key-value store and internally does not have a concept of directories. In some cases, a directory may actually be an empty file, in which case you should use s3_get_meta.

Base.Filesystem.mkdir — Function

mkdir(fp::S3Path; recursive=false, exist_ok=false)

Create an empty directory at the S3 path fp. If recursive, this will create any previously non-existent directories which would contain fp. An error will be thrown if an object exists at fp unless exist_ok.

Note that empty directories in S3 are actually 0-byte objects with the naming convention of a directory.

This will not create a bucket.

Base.read — Function

read(fp::S3Path; byte_range=nothing)

Fetch data from the S3 path as a Vector{UInt8}. A subset of the object can be specified with byte_range which should be a contiguous integer range, e.g. 1:4.

AWSS3.get_config — Function

get_config(fp::S3Path)

Returns the AWS configuration used by the path fp. This can be stored within the path itself, but if not it will be fetched with AWS.global_aws_config().

Internal

AWSS3._s3_exists_dir — Function

_s3_exists_dir(aws::AbstractAWSConfig, bucket, path)

Internal, called by s3_exists.

Checks whether the given directory exists. This is a bit subtle because of how the AWS API handles empty directories. Empty directories are really just 0-byte nodes which are named like directories, i.e. their name has a trailing "/".

What this function does is, given a directory name dir/, check for all keys which are lexographically greater than dir.. The reason for this is that, if dir/ is a 0-byte node, checking for it directly will not reveal its existence due to some rather peculiar design choices on the part of the S3 developers.

In all such cases, if the directory exists it will be seen in the first item returned from S3.list_objects_v2: for empty directories this is because using start-after explicitly excludes dir. itself and dir/ is next; for directories with actual keys, it is guaranteed that the first contained key will start with the directory name.

AWSS3.s3_exists_versioned — Function

s3_exists_versioned([::AbstractAWSConfig], bucket, path, version)

Check if the version specified by version of the object in bucket bucket exists at path.

Note that this function relies on error catching and may be less performant than s3_exists_unversioned which is preferred. The reason for this is that support for versioning in the AWS API is very limited.

AWSS3.s3_exists_unversioned — Function

s3_exists_unversioned([::AbstractAWSConfig], bucket, path)

Returns a boolean whether an object exists at path in bucket.

See s3_exists_versioned to check for specific versions.